third_party/WebKit/LayoutTests/fast/dom/gc-attribute-node-expected.txt

Tests that attribute node wrappers are not prematurely garbage collected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS a.prop is "set"
PASS a.prop is "set"
PASS successfullyParsed is true

TEST COMPLETE
